fix(bot): store max_spendable_freudpoints per server instead of per user
