+++ models/config.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from sqlalchemy import Column, BigInteger, Integer, Text, select
 from sqlalchemy.schema import FetchedValue
-from sqlalchemy.orm import Query
+from sqlalchemy.orm import Query, validates
 
 from discord import Guild
 
@@ -32,9 +32,17 @@ class Config(Base, Model):
     invalid_code_message = Column(Text, FetchedValue(), nullable=False)
     already_verified_message = Column(Text, FetchedValue(), nullable=False)
     welcome_message = Column(Text, FetchedValue(), nullable=False)
+    max_spendable_freudpoints = Column(Integer, nullable=False)
 
     def __repr__(self) -> str:
         return f"Config(guild_id={self.guild_id}, verified_role={self.verified_role}, admin_role={self.admin_role}, logging_channel={self.logging_channel}, confession_approval_channel={self.confession_approval_channel}, confession_channel={self.confession_channel})"
+
+    @validates("max_spendable_freudpoints")
+    def validate_max_spendable_fp_positive(self, key, value):
+        if value < 0:
+            raise ValueError(f"maximum spendable FreudPoints must be at least 0")
+
+        return value
 
     @classmethod
     async def get(cls, guild_id: int) -> Optional["Config"]:
